+++ codegen/tests/test_functions.rs
@@ -1,6 +1,6 @@
 use rhai::module_resolvers::*;
 use rhai::plugin::*;
-use rhai::{Engine, EvalAltResult, Module, RegisterFn, FLOAT};
+use rhai::{Engine, EvalAltResult, Module, FLOAT};
 
 pub mod raw_fn {
     use rhai::plugin::*;