+++ tests/constants.rs
@@ -1,4 +1,4 @@
-use rhai::{Engine, EvalAltResult, ParseErrorType, RegisterFn, Scope, INT};
+use rhai::{Engine, EvalAltResult, ParseErrorType, Scope, INT};
 
 #[test]
 fn test_constant() -> Result<(), Box<EvalAltResult>> {